+++ src/core/Core.js
@@ -38,6 +38,7 @@ const Discord = require("discord.js");
 const CommandScope = require("../util/commands/CommandScope").default;
 const AliasCommand = require("./commands/AliasCommand");
 const { default: BotStatsManager, MESSAGES_TODAY, COMMANDS_EXECUTED } = require("./managers/BotStatsManager");
+const { default: GuildStatsManager } = require("./managers/GuildStatsManager");
 const Translation = require("../modules/i18n/Translation").default;
 
 const BotListManager = require("./managers/BotListManager").default;
@@ -81,12 +82,13 @@ class Core {
         this.locale = new LocaleManager(this.client, this.db);
 
         this.bot_stats = new BotStatsManager(this.db);
+        this.guild_stats = new GuildStatsManager(this.db);
 
-        this.processor = new CommandProcessor(this.client, this.config, this.locale, this.db, this.bot_stats);
-        this.website = new WebsiteManager(this.processor.REGISTRY, this.client, this.config, this.locale, this.db);
+        this.processor = new CommandProcessor(this.client, this.config, this.locale, this.bot_stats, this.guild_stats, this.db);
+        this.website = new WebsiteManager(this.processor.REGISTRY, this.client, this.config, this.locale, this.guild_stats, this.db);
         this.upvotes = new UpvotesManager(this.client, this.db);
 
-        this.member_log = new MemberLog(this.client, this.config, this.locale, this.bot_stats);
+        this.member_log = new MemberLog(this.client, this.config, this.locale, this.bot_stats, this.guild_stats);
 
         this.botlist = new BotListManager(this.client);
         this.presence_status = new PresenceStatusManager(this.client, this.db);
@@ -129,7 +131,7 @@ class Core {
         const install_opts = {
             client: this.client,
             config: this.config, locale: this.locale, db: this.db, error_cases: this.processor.error_cases,
-            presence_status: this.presence_status, bot_stats: this.bot_stats,
+            presence_status: this.presence_status, bot_stats: this.bot_stats, guild_stats: this.guild_stats,
         };
         await Promise.all(files.map(async file => {
             log("%s:", file, "installing...");